refactor: ChatMessageSendResponseDto를 이벤트 Payload로 네이밍 변경 - #18
