+++ chat/src/main/java/com/back/chat/event/ChatMessageSavedEvent.java
@@ -1,9 +1,9 @@
 package com.back.chat.event;
 
 
-import com.back.chat.dto.response.ChatMessageSendResponseDto;
+import com.back.chat.event.payload.ChatMessagePayload;
 
 public record ChatMessageSavedEvent(
         Long roomId,
-        ChatMessageSendResponseDto payload
+        ChatMessagePayload payload
 ) {}
